// app/shared/models/GrievanceType.model.ts
export class GrievanceTypeModel {
     nGrievanceTypeId: number = 0;
     tGrievanceType: string = null;
     nGrievanceTypeOrder: number = 0;
     bIsActive: boolean = false;
}
export class GrievanceTypeViewModel 
{
     nGrievanceTypeId: number = 0;
     tGrievanceType: string="";
     nGrievanceTypeOrder: number = 0;
     dCreatedDt : string = ""
     bIsActive: boolean = false;
}